+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.11 (Ongoing)
+## 0.12 (Ongoing)
+
+* SciptEngine target finding now supports tutoring cards from deck
+* SciptEngine target finding now supports affecting multiple cards from the board by filtering their properties
+
+## 0.11
 
 * Added new card back
 * Made the card back glow-pulse
@@ -9,21 +14,37 @@
 * Cards in Piles are now places face-up or facedown depending on the pile config
 * Card definitions can now be created using json
 * Supports multiple sets of card and script definitions
-* Added Scripting Engine!
+* Added [Scripting Engine](https://github.com/db0/godot-card-gaming/wiki/ScriptDefinitions)!
   * Tasks supported:
     * Rotate Card
     * Flip card face-up/down
     * Move card to a CardContainer
     * Move card to board
     * Move card from Container to Container
+    * Spawn new card
+	* Shuffle CardContainer
+	* Attach to card
+	* Attach card to self
   * Trigger-based execution supported:
-    * card_rotated
-	* card_flipped
-	* card_viewed
-	* card_moved_to_board
-	* card_moved_to_hand
-	* card_moved_to_pile
-	* card_tokens_modified
+    * Card rotated
+	* Card flipped
+	* Card viewed
+	* Card moved to board
+	* Card moved to hand
+	* Card moved to pile
+	* Card_tokens_modified
+	* Card attached to another
+	* Card unattached from another
+  * Filtering of triggers supported:
+	* Card properties
+	* Card rotation degrees
+	* Card facing
+	* Source container of move
+	* Destination container of move
+	* Token count
+	* Token increase/decrease
+	* Token name
+* Scripting also support setting some tasks as costs required and targetting cards on the table.
 
 
 In case you're wondering why the FocusHighlight node has two children panel nodes, instead of making it itself a Panel node.